+++ api_req_air_corps/set_plane/response.ts
@@ -9,15 +9,20 @@
  * type for API response at /kcsapi/api_req_air_corps/set_plane
  */
 export interface APIReqAirCorpsSetPlaneResponse {
-  api_after_bauxite: number
-  api_distance: number
+  api_after_bauxite?: number
+  api_distance: APIDistance
   api_plane_info: APIPlaneInfo[]
 }
 
+export interface APIDistance {
+  api_base: number
+  api_bonus: number
+}
+
 export interface APIPlaneInfo {
-  api_cond: number
-  api_count: number
-  api_max_count: number
+  api_cond?: number
+  api_count?: number
+  api_max_count?: number
   api_slotid: number
   api_squadron_id: number
   api_state: number