+++ api_req_member/get_incentive/response.ts
@@ -10,4 +10,12 @@
  */
 export interface APIReqMemberGetIncentiveResponse {
   api_count: number
+  api_item?: APIItem[]
+}
+
+export interface APIItem {
+  api_getmes: string
+  api_mode: number
+  api_mst_id: number
+  api_type: number
 }