+++ api_req_sortie/ld_airbattle/response.ts
@@ -9,6 +9,7 @@
  * type for API response at /kcsapi/api_req_sortie/ld_airbattle
  */
 export interface APIReqSortieLdAirbattleResponse {
+  api_air_base_attack?: APIAirBaseAttack[]
   api_deck_id: number
   api_eParam: number[][]
   api_eSlot: number[][]
@@ -26,11 +27,19 @@ export interface APIReqSortieLdAirbattleResponse {
   api_stage_flag: number[]
 }
 
-export interface APIKouku {
-  api_plane_from: number[][]
+export interface APIAirBaseAttack {
+  api_base_id: number
+  api_plane_from: Array<number[] | null>
+  api_squadron_plane: APISquadronPlane[]
   api_stage1: APIStage1
-  api_stage2: APIStage2
-  api_stage3: APIStage3
+  api_stage2: APIStage2 | null
+  api_stage3: APIAirBaseAttackAPIStage3 | null
+  api_stage_flag: number[]
+}
+
+export interface APISquadronPlane {
+  api_count: number
+  api_mst_id: number
 }
 
 export interface APIStage1 {
@@ -43,13 +52,34 @@ export interface APIStage1 {
 }
 
 export interface APIStage2 {
+  api_air_fire?: APIAirFire
   api_e_count: number
   api_e_lostcount: number
   api_f_count: number
   api_f_lostcount: number
 }
 
-export interface APIStage3 {
+export interface APIAirFire {
+  api_idx: number
+  api_kind: number
+  api_use_items: number[]
+}
+
+export interface APIAirBaseAttackAPIStage3 {
+  api_ebak_flag: number[]
+  api_ecl_flag: number[]
+  api_edam: number[]
+  api_erai_flag: number[]
+}
+
+export interface APIKouku {
+  api_plane_from: Array<number[] | null>
+  api_stage1: APIStage1
+  api_stage2: APIStage2
+  api_stage3: APIKoukuAPIStage3 | null
+}
+
+export interface APIKoukuAPIStage3 {
   api_ebak_flag: number[]
   api_ecl_flag: number[]
   api_edam: number[]