+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 20180112
+
+recent update 26th Dec
+
+- `api_distance`: must be `Object` for now
+- `api_get_member/mapinfo/response`: `api_event_map` structure change
+- `/kcsapi/api_get_member/picture_book`: fields added for item picture
+- `/kcsapi/api_req_sortie/ld_shooting`: new API
+- `/kcsapi/api_req_combined_battle/ld_shooting`: a new API but no response has been captured yet
+
+and various minor typing updates
+
 ## 20181202
 
 recent update: 29th Nov